requirements.txt
azure-functions
textblob
